fixed update products

+++ routes/api/product-routes.js
@@ -1,7 +1,6 @@
 // Imports express router and models
 const router = require('express').Router();
 const { Product, Category, Tag, ProductTag } = require('../../models');
-const { findByPk } = require('../../models/Product');
 
 // GET route for `/api/product`
 router.get('/', async (req, res) => {
@@ -75,32 +74,36 @@ router.put('/:id', (req, res) => {
   })
     .then((product) => {
       // Finds all associated tags from ProductTag
+      
       return ProductTag.findAll({ where: { product_id: req.params.id } });
     })
     .then((productTags) => {
+      if (req.body.tagIds && req.body.tagIds.length) {
       // Gets list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
+        const productTagIds = productTags.map(({ tag_id }) => tag_id);
       // Creates a filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
-      // Figures out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+        const newProductTags = req.body.tagIds
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              product_id: req.params.id,
+              tag_id,
+            };
+          });
+        // Figures out which ones to remove
+        const productTagsToRemove = productTags
+         .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+         .map(({ id }) => id);
 
-      // Runs both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+        // Runs both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags),
+        ]);
+      }
+      res.status(200).json(productTags);
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
+    .then((updatedProductTags) => res.status(200).json(updatedProductTags))
     .catch((err) => {
       res.status(400).json(err);
     });
